refactor(countries): move all country-related files into components/Countries
